Simplify conversion function
